refactor: extract useCloudSync hook
Moves cloudSyncConfig, cloudSyncStatus, cloudSyncError, cloudSyncLastSynced,
cloudSyncConflict state and all cloud-sync refs (cloudSyncDebounceRef,
suppressCloudUploadRef, suppressTimestampRef, suppressClearPendingRef,
cloudSyncInProgressRef, cloudSyncInitialDoneRef, cloudSyncDownloadRef,
cloudSyncErrorCountRef, cloudSyncBackoffUntilRef) plus the config persist
effect into a dedicated useCloudSync hook.
